+++ CHANGELOG.md
@@ -25,6 +25,7 @@ This new version contains significant breaking changes to the graph database and
 - Added the ability to call methods on Intent's via command line intents with `#m:`
 - Added `L` to the diff UI for logcat strings
 - Added `Message` types to the Parcel mini language
+- Added `find outgoing-calls`
 
 # 4.2.0